Conflicts:
	Source/UI/SDNumberTextField.m

+++ Source/UI/SDTextFieldPicker.m
@@ -92,7 +92,7 @@
 
 - (void)setPickerItems:(NSArray<NSString> *)pickerItems
 {
-    [_pickerButton configureAsItemPicker:pickerItems completion:^(BOOL canceled, int selectedItemIndex, NSString *selectedItem) {
+    [_pickerButton configureAsItemPicker:pickerItems completion:^(BOOL canceled, NSInteger selectedItemIndex, NSString *selectedItem) {
         NSString *otherString = [selectedItem lowercaseString];
         
         // if they selected "Other" or hit cancel, then go back to the text field.
